Added another fix for alignment issues (specifically, for Asian locales)
The fixed spacer didn't account for the Asian font size, which is
larger-than-normal and lead to the text being written over the bar
displays. Now, this is no longer an issue, although the space "appears"
somewhat too big for those locales compared to how it looks when using a
smaller font.