GUI: Removed unnecessary condition to allow updating as combat is initialised
This fixes #16 since now the command gets to hide the window just as
combat is being entered, but without being affected by combat lockdown.

There don't appear to be any taint issues arising from it; either, but
this might require some more in-depth testing with the taintlog enabled
to be sure.