Fixed miscalculation of available traits when starting at artifact tier 1
The API would not account for the fact that tiers update automatically
now. Since the costs for <=35 traits are identical for tier 1 and 2,
manually setting it to 2 will fix this.

Obviously this is more of a hack and not a real solution, which would
require much more effort. However, there is no downside to this at the
moment, and if Antorus brings artifact tier 3 it can easily be updated
to use that instead (unless it requires manually unlocking it once
more).